Реализована поддержка карт нормалей.

+++ Source/Code/Engine/Render/RenderSystem.h
@@ -111,7 +111,7 @@ inline Color operator/(const Color& color, const float Scalar)
 
 inline float DistanceBetweenColor(const Color& Color1, const Color& Color2)
 {
-	return powf((float)Color1.R - (float)Color2.R, 2.0f) + powf((float)Color1.G - (float)Color2.G, 2.0f) + powf((float)Color1.B - (float)Color2.B, 2.0f);
+	return powf(Color1.R - Color2.R, 2.0f) + powf(Color1.G - Color2.G, 2.0f) + powf(Color1.B - Color2.B, 2.0f);
 }
 
 #define SAFE_DX(Func) CheckDXCallResult(Func, u#Func);
